chore: Refactor registry tests (#4223)
chore: refactor registry tests
